requirements.txt
keras==2.10.0
numpy==1.24.2
pandas==1.5.3
python-telegram-bot==20.2
tensorflow==2.10.1
